requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
jsonschema>=4.21.1
pillow>=10.2.0
python-telegram-bot>=20.8
pytz>=2024.1
pytest>=7.4.4
pytest-asyncio>=0.23.3
pytest-cov>=4.1.0
pytest-mock>=3.12.0 